+++ wordpress/phpunit.xml
@@ -8,7 +8,6 @@
         <testsuite name="Test Suite">
             <directory suffix="Test.php">./tests</directory>
             <directory suffix="Test.php">./wp-content/themes/cds-default/tests</directory>
-            <directory suffix="Test.php">./wp-content/mu-plugins/cds-base/classes/tests</directory>
             <directory suffix="Test.php">./wp-content/mu-plugins/cds-base/classes/Modules/*/tests</directory>
         </testsuite>
     </testsuites>
