+++ wordpress/wp-content/mu-plugins/cds-base/classes/Modules/EncryptedOption/EncryptedOption.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CDS;
+namespace CDS\Modules\EncryptedOption;
 
 use Illuminate\Encryption\Encrypter;
 use Illuminate\Support\Str;
