+++ wordpress/wp-content/mu-plugins/cds-base/classes/Setup.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace CDS;
 
+use CDS\Modules\EncryptedOption\EncryptedOption;
 use CDS\Modules\Blocks\Blocks;
 use CDS\Modules\Cleanup\AdminBar as CleanupAdminBar;
 use CDS\Modules\Cleanup\AdminStyles as CleanupAdminStyles;
